+++ oalg-base/src/OAlg/Entity/Sequence/PSequence.hs
@@ -1,7 +1,7 @@
 
 {-# LANGUAGE NoImplicitPrelude #-}
 
-{-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE TypeFamilies, TypeOperators #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE FlexibleContexts #-}
@@ -17,11 +17,11 @@
 -- partially defined sequences of items in @__x__@ with a totally ordered index type @__i__@.
 module OAlg.Entity.Sequence.PSequence
   ( -- * Sequence
-    PSequence(..), iProxy
+    PSequence(..), psqxs, iProxy
   , psqSpan
-  , psqEmpty, psqIsEmpty, psqxs, psequence
+  , psqEmpty, psqIsEmpty, psequence
   , psqHead, psqTail
-  , psqMap, psqMapShift
+  , psqMap, psqMapShift, psqMapWithIndex, Monotone(..)
   , psqFilter
   , psqSplitWhile
   , psqInterlace
@@ -29,12 +29,21 @@ module OAlg.Entity.Sequence.PSequence
   , psqAppend
   , psqShear
   , psqSwap
+  , psqTree
+  , psqFromTree
 
     -- * Tree
-  , PTree(..),psqTreeLookup, psqTree, psqFromTree, psqTreeMax, psqTreeMin, psqTreeSupport
+  , PTree(..), ptrxs, ptrx
+  , ptrEmpty
+  , ptrMax, ptrMin, ptrSpan
+  , ptrFilter, ptrFilterWithIndex
+  , ptrMap, ptrMapShift, ptrMapWithIndex
   
     -- * X
   , xPSequence
+
+    -- * Proposition
+  , prpPSequence, prpPTreeFilter
   ) where
 
 import Control.Monad 
@@ -173,20 +182,45 @@ psqIsEmpty (PSequence []) = True
 psqIsEmpty _              = False
 
 --------------------------------------------------------------------------------
--- psqMap -
+-- Monotone -
 
--- | maps the entries, where the indices are preserved.
-psqMap :: (x -> y) -> PSequence i x -> PSequence i y
-psqMap f (PSequence xis) = PSequence $ map (\(x,i) -> (f x,i)) xis
 
-instance Functor (PSequence i) where fmap = psqMap
+-- | predicate for strict monoton mappings.
+--
+-- __Property__ Let @_i__@, @__j__@ two 'Ord'-types, and @'Monotone' f@ in @'Monotone' __i__ __j__@, then
+-- holds: For all @x@, @y@ in @__i__@ holds: @'compare' (f x) (f y) '==' 'compare' x y@. 
+newtype Monotone i j = Monotone (i -> j)
+
+instance (Ord i, Ord j, XStandard i, Show i) => Validable (Monotone i j) where
+  valid (Monotone f) = Label "Monotone"
+    :<=>: Forall xy (\(x,y)
+                     -> ((compare (f x) (f y) == compare x y) :?> Params ["x":=show x,"y":=show y])
+                    )
+    where xy = xTupple2 xStandard xStandard
+
+--------------------------------------------------------------------------------
+-- psqMapWithIndex -
+
+-- | maps of a 'PSequence' according to the given strict monotone mapping and the given indexed
+-- mapping.  
+psqMapWithIndex :: Monotone i j -> ((x,i) -> y) -> PSequence i x -> PSequence j y
+psqMapWithIndex (Monotone j) y (PSequence xis) = PSequence $ map yj xis where yj xi = (y xi,j $ snd xi)
   
 --------------------------------------------------------------------------------
 -- psqMapShift -
 
--- | maps and shifts a partial sequence.
+-- | shifts the indices of a partial sequence.
 psqMapShift :: Number i => i -> ((x,i) -> y) -> PSequence i x -> PSequence i y
-psqMapShift o f (PSequence xs) = PSequence $ map f' xs where f' xi = (f xi,snd xi + o)
+psqMapShift t = psqMapWithIndex (Monotone (+t))
+
+--------------------------------------------------------------------------------
+-- psqMap -
+
+-- | maps the entries, where the indices are preserved.
+psqMap :: (x -> y) -> PSequence i x -> PSequence i y
+psqMap f = psqMapWithIndex (Monotone id) (f . fst)
+
+instance Functor (PSequence i) where fmap = psqMap
 
 --------------------------------------------------------------------------------
 -- PSequence - Entity -
@@ -369,7 +403,7 @@ psqSwap k l = psqShear (sk,k) (sl,l) where
 --------------------------------------------------------------------------------
 -- xPSequence -
 
--- | @'xPSequence' n m@ random variable of partially defined sequences with maximal length
+-- | @'xPSequence' n m@ random variable of partially defined sequences with maximal length of
 --   @'min' n m@.
 xPSequence :: Ord i => N -> N -> X x -> X i -> X (PSequence i x)
 xPSequence n m xx xi = do
@@ -382,12 +416,6 @@ xPSequence n m xx xi = do
 
 -- | binary tree for efficient retrieving elements of a partially defined sequence.
 newtype PTree i x = PTree (Maybe (Tree i (x,i))) deriving (Show,Eq,Ord)
-
-instance Functor (PTree i) where
-  fmap _ (PTree Nothing)  = PTree Nothing
-  fmap f (PTree (Just t)) = PTree $ Just $ tmap f t where
-    tmap f (Leaf (x,i)) = Leaf (f x,i)
-    tmap f (Node i l r) = Node i (tmap f l) (tmap f r)
 
 instance Foldable (PTree i) where
   foldMap _ (PTree Nothing)  = mempty
@@ -416,7 +444,14 @@ instance (Entity x, Entity i, Ord i) => Validable (PTree i x) where
         Node i' l r -> valid i' && vldl i' l && vldr i' r
     
 instance (Entity x, Entity i, Ord i) => Entity (PTree i x)
-      
+
+--------------------------------------------------------------------------------
+-- ptrEmpty -
+
+-- | the empty 'PTree'.
+ptrEmpty :: PTree i x
+ptrEmpty = PTree Nothing
+
 --------------------------------------------------------------------------------
 -- psqTree -
 
@@ -437,36 +472,127 @@ psqFromTree (PTree Nothing)  = psqEmpty
 psqFromTree (PTree (Just t)) = PSequence $ toList t
 
 --------------------------------------------------------------------------------
--- psqTreeLookup -
+-- ptrxs -
 
--- | retrieving a value from the tree.
-psqTreeLookup :: Ord i => PTree i x -> i -> Maybe x
-psqTreeLookup (PTree Nothing)  _ = Nothing
-psqTreeLookup (PTree (Just t)) i = if i' == i then Just x else Nothing where (x,i') = lookup t i
+-- | the underlying list of indexed values.
+ptrxs :: PTree i x -> [(x,i)]
+ptrxs = psqxs . psqFromTree
 
 --------------------------------------------------------------------------------
--- psqTreeMin -
+-- ptrx -
+
+-- | retrieving a value from the tree.
+ptrx :: Ord i => PTree i x -> i -> Maybe x
+ptrx (PTree Nothing)  _ = Nothing
+ptrx (PTree (Just t)) i = if i' == i then Just x else Nothing where (x,i') = trLookup t i
+
+--------------------------------------------------------------------------------
+-- ptrMin -
 
 -- | the minimal index.
-psqTreeMin :: PTree i x -> Closure i
-psqTreeMin (PTree Nothing)  = PosInf
-psqTreeMin (PTree (Just t)) = pmin t where
+ptrMin :: PTree i x -> Closure i
+ptrMin (PTree Nothing)  = PosInf
+ptrMin (PTree (Just t)) = pmin t where
   pmin (Leaf (_,i)) = It i
   pmin (Node _ l _) = pmin l
 
 --------------------------------------------------------------------------------
--- psqTreeMax -
+-- ptrMax -
 
 -- | the maximal index.
-psqTreeMax :: PTree i x -> Closure i
-psqTreeMax (PTree Nothing)  = NegInf
-psqTreeMax (PTree (Just t)) = pmax t where
+ptrMax :: PTree i x -> Closure i
+ptrMax (PTree Nothing)  = NegInf
+ptrMax (PTree (Just t)) = pmax t where
   pmax (Leaf (_,i)) = It i
   pmax (Node _ _ r) = pmax r
 
 --------------------------------------------------------------------------------
--- psqTreeSupport -
+-- ptrSpan -
 
--- | the support, i.e the minimal and the maximal index of the tree.
-psqTreeSupport :: PTree i x -> (Closure i,Closure i)
-psqTreeSupport t = (psqTreeMin t,psqTreeMax t)
+-- | the span, i.e the minimal and the maximal index of the tree.
+ptrSpan :: PTree i x -> Span i
+ptrSpan t = (ptrMin t,ptrMax t)
+
+--------------------------------------------------------------------------------
+-- ptrFilter -
+
+-- | the sub tree containing all leafs satisfying the given predicate.
+ptrFilter :: (x -> Bool) -> PTree i x -> PTree i x
+ptrFilter p (PTree t) = PTree (t >>= trFilter (p . fst))
+
+--------------------------------------------------------------------------------
+-- ptrFilterWithIndex -
+
+-- | the sub tree containing all leafs satisfying the given indexed predicate.
+ptrFilterWithIndex :: ((x,i) -> Bool) -> PTree i x -> PTree i x
+ptrFilterWithIndex p (PTree t) = PTree (t >>= trFilter p)
+
+--------------------------------------------------------------------------------
+-- ptrMapWithIndex -
+
+-- | maps a 'PTree' according to the given strict monotone mapping and the given indexed
+-- mapping.
+ptrMapWithIndex :: Monotone i j -> ((x,i) -> y) -> PTree i x -> PTree j y
+ptrMapWithIndex _ _ (PTree Nothing)           = PTree Nothing
+ptrMapWithIndex (Monotone j) y (PTree (Just t)) = PTree $ Just $ mapwi t where
+  mapwi (Leaf xi)    = Leaf (y xi,j $ snd xi)
+  mapwi (Node i l r) = Node (j i) (mapwi l) (mapwi r)
+
+--------------------------------------------------------------------------------
+-- ptrMapShift -
+
+-- | shifts the indices of a 'PTree'.
+ptrMapShift :: Number i => i -> ((x,i) -> y) -> PTree i x -> PTree i y
+ptrMapShift t = ptrMapWithIndex (Monotone (+t))
+
+--------------------------------------------------------------------------------
+-- ptrMap -
+
+-- | maps a 'PTree' according to the given mapping.
+ptrMap :: (x -> y) -> PTree i x -> PTree i y
+ptrMap f = ptrMapWithIndex (Monotone id) (f . fst)
+
+instance Functor (PTree i) where fmap = ptrMap
+
+--------------------------------------------------------------------------------
+-- prpTreeFilter -
+
+-- | validates the function 'pspTreeFilter'
+prpPTreeFilter :: N -> Statement
+prpPTreeFilter n = Prp ("PTreeFilter " ++ show n) :<=>:
+  Forall xps (\pxs -> (  (filter (p . fst) $ psqxs pxs)
+                      == (psqxs $ psqFromTree $ ptrFilter p $ psqTree pxs)
+                      ) :?> Params ["pxs":=show pxs]
+             ) where
+
+  xps = xOneOfXW [(1,return psqEmpty),(100,xPSequence n n (xZB (-100) 100) xN)]
+  p x = x `mod` 2 == 0
+
+--------------------------------------------------------------------------------
+-- prpMapShiftPSequence -
+
+-- | validates the shifting of a 'PSequence' and 'PTree'.
+prpMapShiftPSequence :: N -> Statement
+prpMapShiftPSequence n = Prp "ShiftPSequencePTree" :<=>:
+  Forall xxitf (\  (xis,t,f)
+               ->  let xis' = ptrMapShift t f $ psqTree xis in
+                     And [ valid t
+                         , (psqMapShift t f xis == psqFromTree xis')
+                             :?> Params ["xis":=show xis,"t":=show t]
+                         ]
+               )
+
+  where xxitf :: X (PSequence Z Z,Z,(Z,Z) -> Z)
+        xxitf = xTupple3 (xPSequence n n xZ xZ) xZ xf
+        xf = xOneOf [uncurry (+),uncurry (-)]
+  
+--------------------------------------------------------------------------------
+-- prpPSequence -
+
+-- | validity of 'PSequence'.
+prpPSequence :: Statement
+prpPSequence = Prp "PSequence"
+  :<=>: And [ prpPTreeFilter 60
+            , prpMapShiftPSequence 76
+            ]
+