+++ oalg-base/src/OAlg/Proposition.hs
@@ -51,6 +51,7 @@ prpOAlgBase = Prp "OAlgBase"
             , Label "Permutation" :<=>: prpPermutation
             , Label "Matrix"  :<=>: prpMatrixZ
             , Label "Vector"  :<=>: prpRepMatrixZ 8 15
+            , prpPSequence
             , prpFSequence
             ]